update: ゲーム開始時の処理を別クラスに分離
単一責務になっていなかったので分離した
