+++ Assets/Prefabs/alphaUI/Canvas/MainCanvas.prefab
@@ -212,6 +212,7 @@ GameObject:
   m_Component:
   - component: {fileID: 5745597649211064877}
   - component: {fileID: 2106186069054512436}
+  - component: {fileID: 1939404891127187098}
   - component: {fileID: 792960411302418462}
   m_Layer: 5
   m_Name: Presenter
@@ -249,6 +250,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 248857c8872140cfb5d7ed6b59f56530, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1939404891127187098
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5775137310579412787}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bb6bad39765d49ca98a5c273520c46f6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &792960411302418462
