+++ Assets/Scripts/InGame/StateMachine/State/IGameStartedPresenter.cs
@@ -1,0 +1,7 @@
+namespace September.Common
+{
+    public interface IGameStartedPresenter
+    {
+        public void OnGameStarted();
+    }
+}
